requirements.txt
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
numpy>=1.23.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
streamlit>=1.33.0
pyyaml>=6.0.1
